platform/backend/requirements.txt
fastapi
uvicorn[standard]
httpx
python-jose[cryptography]
psycopg2-binary
python-dotenv
